Added solution for Day 14 Puzzle 2.

+++ AdventOfCode/Chemistry/Chemical.cs
@@ -2,13 +2,13 @@
 {
     public class Chemical
     {
-        public Chemical(string name, int quantity)
+        public Chemical(string name, long quantity)
         {
             this.Name = name;
             this.Quantity = quantity;
         }
 
         public string Name { get; }
-        public int Quantity { get; }
+        public long Quantity { get; }
     }
 }
